workaround to keep stream thumbs updated
I tried having the thumb= to a Callback, which did a get request for the
image then returned a DataObject. That didn't cause the cache to reset.
The only way to update the cache is to change the filename to something
different.

So I'm just appending "#timestamp" to the jpg url for the preview
images.