favourite games menu
adds a favourite games preference item, and a fav games menu.

The value in the preferences should be a comma separated list of game
names. They shouldn't need to be exact matches because it's using the
search streams of the twitch api.

"hearthstone,diablo, league of legends" should be a valid string, that
adds 3 items to the fav games menu.