Added support to watch uploads
Twitch added a possibilty to upload videos a while ago. VODs can now
either be of type highlight, archive or upload. In order to be able
to see them the query must use the new 'broadcast_type' parameter.

This plugin will now display all uploads of a channel in a sepate list
similar to how it was done so far for past broadcasts or highlights.